+++ cluster/pulumi/canton-network/src/cometbft.ts
@@ -1,12 +1,14 @@
 import { Output } from '@pulumi/pulumi';
 
-import { CLUSTER_DNS_NAME, CLUSTER_NAME, ExactNamespace, installCNHelmChart } from './utils';
+import { CLUSTER_BASENAME, CLUSTER_DNS_NAME, ExactNamespace, installCNHelmChart } from './utils';
 
 const sv1NodeConfig = {
   id: '5af57aa83abcec085c949323ed8538108757be9c',
   privateKey:
     '/7L74Bs18740fTPdEL04BeO2Gs+1lzEeCjAiB1DYcysmLnU1FAkg/Ho9XsOiIp4U/KT/YNrtIi/A0prm/Ew3eQ==',
-  externalAddress: p2pExternalAddress('sv-1'),
+  identifier: 'cometbft-sv-1',
+  externalAddress: p2pExternalAddress('cometbft-sv-1', 26656),
+  istioPort: 26656,
   validator: {
     keyAddress: '8A931AB5F957B8331BDEF3A0A081BD9F017A777F',
     privateKey:
@@ -19,7 +21,7 @@ const founder = {
   nodeId: sv1NodeConfig.id,
   publicKey: sv1NodeConfig.validator.publicKey,
   keyAddress: sv1NodeConfig.validator.keyAddress,
-  externalAddress: p2pServiceAddress('sv-1'),
+  externalAddress: p2pServiceAddress('cometbft-sv-1', 'sv-1'),
 };
 
 const nodeConfigs: {
@@ -30,7 +32,9 @@ const nodeConfigs: {
     id: 'c36b3bbd969d993ba0b4809d1f587a3a341f22c1',
     privateKey:
       '1Je33z2g+Dj2UWLqnsO+xwUwbalIS0LLcYAoj+fYuEE2le4kJjJ0h+L7FfVg+3mbgvrikdke91I2X5C2frj0Eg==',
-    externalAddress: p2pExternalAddress('sv-2'),
+    identifier: 'cometbft-sv-2',
+    externalAddress: p2pExternalAddress('cometbft-sv-2', 26666),
+    istioPort: 26666,
     validator: {
       keyAddress: '04A57312179F1E0C93B868779EE4C7FAC41666F0',
       privateKey:
@@ -42,7 +46,9 @@ const nodeConfigs: {
     id: '0d8e87c54d199e85548ccec123c9d92966ec458c',
     privateKey:
       'DdbW/buPo4TXxW+/cvQxp5Lh1BZyH5GYHGoU0uTUQA/S1oZ32DDu1+CZhtrZhqEFMuxPlXVXvyvXsZLBdCkgdQ==',
-    externalAddress: p2pExternalAddress('sv-3'),
+    identifier: 'cometbft-sv-3',
+    externalAddress: p2pExternalAddress('cometbft-sv-3', 26676),
+    istioPort: 26676,
     validator: {
       keyAddress: 'FFF137F42421B0257CDC8B2E41F777B81A081E80',
       privateKey:
@@ -54,7 +60,9 @@ const nodeConfigs: {
     id: 'ee738517c030b42c3ff626d9f80b41dfc4b1a3b8',
     privateKey:
       'xMB8gnYbacyZqU94cgwJBK2OJO3DffO12uHgeieotVj/Q9LbZEwLue9GnG8+G5GNRDgX8z75txr/Z541Uqyb3A==',
-    externalAddress: p2pExternalAddress('sv-4'),
+    identifier: 'cometbft-sv-4',
+    externalAddress: p2pExternalAddress('cometbft-sv-4', 26686),
+    istioPort: 26686,
     validator: {
       keyAddress: 'DE36D23DE022948A11200ABB9EE07F049D17D903',
       privateKey:
@@ -64,19 +72,28 @@ const nodeConfigs: {
   },
 };
 
+/**
+ * The CometBft deployment uses a different port for the istio VirtualService for each node
+ * Then all the ports must be added to the gateway so that we can forward the traffic as expected.
+ * This is done because CometBft does not actually support adding multiple nodes with the same ip:port configuration.
+ * It seems that CometBft stores the address of known peers by actually storing the IP:Port combination and discarding the used dns,
+ * therefore having only different DNS entries that point to a different service is not enough.
+ * Furthermore, even if we register multiple istio VirtualServices with different hosts, but for the same port in the gateway,
+ * istio will just ignore the host criteria for TCP ports.
+ * */
 export function installCometBftNode(
   xns: ExactNamespace,
   nodename: string,
   onboardingName: string
 ): void {
   installCNHelmChart(xns, nodename + '-cometbft', 'cn-cometbft', {
-    svNodeId: nodename,
     nodeName: onboardingName,
     imageName: 'cometbft',
     founder: founder,
     istioVirtualService: {
       enabled: true,
       gateway: 'cluster-ingress/apps-gateway',
+      port: nodeConfigs[nodename].istioPort,
     },
     node: nodeConfigs[nodename],
     peers: Object.keys(nodeConfigs)
@@ -90,11 +107,11 @@ export function installCometBftNode(
          * */
         return {
           nodeId: nodeConfigs[svName].id,
-          externalAddress: p2pServiceAddress(svName),
+          externalAddress: p2pServiceAddress(`cometbft-${svName}`, svName),
         };
       }),
     genesis: {
-      chainId: CLUSTER_NAME,
+      chainId: `${CLUSTER_BASENAME}.canton`,
     },
   });
 }
@@ -106,14 +123,16 @@ type NodeConfig = {
     privateKey: Output<string> | string;
     publicKey: Output<string> | string;
   };
+  istioPort: number;
   externalAddress: string;
+  identifier: string;
   id: Output<string> | string;
 };
 
-function p2pExternalAddress(nodename: string): string {
-  return `${nodename}.svc.${CLUSTER_DNS_NAME}:26656`;
+function p2pExternalAddress(nodename: string, port: number): string {
+  return `${nodename}.svc.${CLUSTER_DNS_NAME}:${port}`;
 }
 
-function p2pServiceAddress(nodename: string): string {
-  return `${nodename}-cometbft-p2p.${nodename}.svc.cluster.local:26656`;
+function p2pServiceAddress(nodename: string, namespace: string): string {
+  return `${nodename}-cometbft-p2p.${namespace}.svc.cluster.local:26656`;
 }